==> src/1.download_data.ipynb <==
# Download Data
Saves non-video data

%load_ext autoreload
%autoreload 2

import os

import re
import google_auth_oauthlib.flow
from googleapiclient.discovery import build
import googleapiclient.errors
from pprint import pprint
import json
from youtube_transcript_api import YouTubeTranscriptApi
from tqdm import tqdm
import youtube_dl
import pickle
from glob import glob
import pandas as pd
import numpy as np
from lib.Episode import Episode, EpisodeFactory
import time
import shutil

ID = "lex"
BASE = f"../data/{ID}"
VIDEOS_LOCATION = f"/Volumes/{ID}/videos/"
WEBSITE = f"../../{ID}-vis/public/"

UPLOADS = BASE + "/uploads.json"
STATS = BASE + "/stats.json"
COMMENTS = BASE + "/comments.json"
CAPTIONS = BASE + "/captions.json"
CAPTIONS_FAILED = BASE + "/captions_failure.json"

# Disable OAuthlib's HTTPS verification when running locally.
# *DO NOT* leave this option enabled in production.
scopes = ["https://www.googleapis.com/auth/youtube.readonly"]
API_KEY = "AIzaSyApvC_9XhjS4Hun7JqqAsUuAL3eXhY8rmM"

os.environ["OAUTHLIB_INSECURE_TRANSPORT"] = "1"

api_service_name = "youtube"
api_version = "v3"

# https://developers.google.com/resources/api-libraries/documentation/youtube/v3/python/latest/
youtube = build(api_service_name, api_version, developerKey=API_KEY)

youtube.channels().list(
    # part="snippet,contentDetails,statistics,uploads",
    part="contentDetails",
    forUsername="lexfridman"
).execute()

UPLOADS_ID = "UUzQUP1qoWDoEbmsQxvdjxgQ" # JRE
UPLOADS_ID = "UUSHZKyawb77ixDdsGog4iWA" # LEX

### Get uploads
def get_video_ids(upload_id):
    page_token = None
    items = []
    total_results = 0

    while True:
        print(f"{len(items)}/{total_results} -> {page_token}", end="\r")

        request = youtube.playlistItems().list(
            part="id,contentDetails",
            playlistId=upload_id,
            pageToken=page_token,
            maxResults=50,
        )
        response = request.execute()
        items.extend([item["contentDetails"]["videoId"] for item in response["items"]])

        total_results = response["pageInfo"]["totalResults"]

        if "nextPageToken" in response:
            page_token = response["nextPageToken"]
        else:
            break

    return items

try:
    video_ids = get_video_ids(UPLOADS_ID)
except Exception as e:
    print(e)

len(video_ids)

def get_uploads():
    uploads = []
    
    for id in tqdm(video_ids): 
        response = youtube.videos().list(
            id=id,
            part="contentDetails,id,liveStreamingDetails,recordingDetails,snippet,statistics,topicDetails,status",
        ).execute()
        
        # Sanity check to ensure all responses have length 1
        if len(response['items']) != 1:
            print("There should not be more than one item for", id)
        
        uploads.append(response['items'][0])
        
    return uploads

uploads = get_uploads()
with open(UPLOADS, "w") as f:
    f.write(json.dumps(uploads))

with open(UPLOADS, "r") as f:
    uploads = json.load(f)
# Contents of uploads
len(uploads), uploads[15]

def get_captions():
    failed_to_get_captions = []
    captions = {}
    with open(CAPTIONS, "r") as f:
        captions = json.loads(f.read())
    with open(UPLOADS, "r") as f:
        uploads = json.loads(f.read())
    
    downloaded = 0
    for upload in tqdm(uploads):
        id = upload["id"]

        if id in captions:
            continue

        try:
            caption = YouTubeTranscriptApi.get_transcript(id)
            captions[id] = caption
            downloaded += 1
        except:
            failed_to_get_captions.append(id)

        # save progress
        if downloaded % 10 == 0:
            with open(CAPTIONS, "w") as f:
                f.write(json.dumps(captions))

    print("Downloaded", downloaded, "items")
    return captions, failed_to_get_captions

captions, failed_to_get_captions = get_captions()

with open(CAPTIONS, "w") as f:
    f.write(json.dumps(captions))

with open(CAPTIONS_FAILED, "w") as f:
    f.write(json.dumps(failed_to_get_captions))
    
print(len(captions.items()), len(failed_to_get_captions))

print(len(captions.items()), len(failed_to_get_captions))

print("Missing the following items")
[[(u['snippet']['title'], u['contentDetails']['duration']) for u in uploads if u['id'] == id][0] for id in failed_to_get_captions]

# Create Episodes cache
Create episode object (lib/Episode) out of each upload

CACHE_ALL = f"./{ID}-episodes.pickle"
CACHE_SMALL = f"./{ID}-episodes-small.pickle"
CACHE = CACHE_ALL

print(f"Generating new pickle {CACHE}...")

factory = EpisodeFactory(BASE)
episodes = factory.create_episodes(skip_comments=True)

with open(CACHE, "wb") as f:
    if CACHE == CACHE_SMALL:
        pickle.dump(episodes[:100], f)
    else:
        pickle.dump(episodes, f)

print(f"Number of loaded episodes: {len(episodes)}")

# Save CSV version of this to website directory
ep_df = pd.DataFrame(
    [
        (e.video_id, e.title, e.number, e.published_at, e.guests, e.is_main_episode, e.likes, e.dislikes, e.views, e.comment_count) 
         for e in episodes
    ],
    columns=["id", "title", "number", "published", "guests", "main", "likes", "dislikes","views","commentCount"]
)

# ep_df.to_csv(WEBSITE + "episodes.csv")
ep_df.head()

# Download Videos

CACHE = f"./{ID}-episodes.pickle"

with open(CACHE, "rb") as f:
    episodes = pickle.load(f)

main_eps = [e for e in episodes if e.is_main_episode]
print("Total main episodes", len(main_eps))

video_files = list(glob(VIDEOS_LOCATION + "*.mp4"))
video_files_ids = [v[-15:-4] for v in video_files]
downloaded_main = [e for e in main_eps if e.video_id in video_files_ids]

# sort to give priority to newer episodes first
missing_videos = sorted(
    list(set(main_eps) - set(downloaded_main)), 
    key=lambda ep: ep.number, 
    reverse=True,
)

print("Total missing main videos", len(missing_videos))

ydl = youtube_dl.YoutubeDL(
    {
        # MP4 at 360p
        "format": "18",
        # "cookiefile": "./youtube-dl-cookies.txt"
    }
)

downloaded = 0
with ydl:
    for ep in missing_videos:
        video = f"http://www.youtube.com/watch?v={ep.video_id}"

        try:
            result = ydl.download([video])
            print(result)
            downloaded += 1
            
            # Couldn't get option "outtml" to work, manually move instead
            try:
                file = f"{ep.title}-{ep.video_id}.mp4"
                shutil.move(f"./{file}", VIDEOS_LOCATION + file)
            except Exception as e:
                print("Could not move", ep, e)
        
        except Exception as e:
            print("could not download", ep.title)
        
        time.sleep(15)

print("Downloaded", downloaded, "videos")

## Unused but may be helpful to someone

def get_comments(videoId, max=1500):
    page_token = None
    items = []
    # print(videoId)

    while True and len(items) < max:
        # print(f"\t{len(items)}/{max}")
        request = youtube.commentThreads().list(
            part="id,snippet",
            videoId=videoId,
            pageToken=page_token,
            maxResults=500,
            order="relevance",
        )
        response = request.execute()
        items.extend(response["items"])

        if "nextPageToken" in response:
            page_token = response["nextPageToken"]
        else:
            break

    return items

def get_comments():
#     Get comments and statistics for each video
    with open(UPLOADS, "r") as f:
        uploads = json.loads(f.read())

        # likeCount, textDisplay, textOriginal, publishedAt
        # authorDisplayName, authorChannelId
        i = 0
        comment_threads = {}
        with open(COMMENTS, "r") as f:
            comment_threads = json.load(f)

        for upload in tqdm.tqdm(uploads):
            videoId = upload["snippet"]["resourceId"]["videoId"]

            if videoId not in comment_threads:
                comments = get_comments(videoId)
                comment_threads[videoId] = comments
                print(videoId, len(comments))
            else:
                print("Already saved thread for", videoId)

            i += 1
            # save progress
            if i % 25 == 0:
                with open(COMMENTS, "w") as f:
                    f.write(json.dumps(comment_threads))

        with open(COMMENTS, "w") as f:
            f.write(json.dumps(comment_threads))






# I use video data instead
def download_audio():
    audio_file = lambda id: f"data/jre/audio/{id}"
    with open(UPLOADS, "r") as f:
        uploads = json.loads(f.read())

        for upload in tqdm(uploads):
            id = upload["contentDetails"]["videoId"]
            yt = None
            if os.path.exists(audio_file(id)):
                print("File exists", id, "skipping...")
                continue

            try:
                yt = YouTube(f"https://www.youtube.com/watch?v={id}")
            except Exception as e:
                print("Could not load api", id, str(e))
                continue

            audio_streams = yt.streams.filter(only_audio=True)

            if len(audio_streams) == 0:
                print("No audio streams available for", id)
                continue

            mp4_streams = [s for s in audio_streams if s.mime_type == "audio/mp4"]
            stream = audio_streams[0] if len(mp4_streams) == 0 else mp4_streams[0]

            print(f"Downloading [{id}] {stream.mime_type} {stream.abr}")
            stream.download(audio_file(id))
            print("\t... Done")

# Unused utils
def get_timestamps(c):
    timestamp = r"(\d?\d?:)?\d?\d?:\d{2}"

    if "topLevelComment" in c["snippet"]:
        snippet = c["snippet"]["topLevelComment"]["snippet"]

        if re.search(timestamp, snippet["textOriginal"]) is None:
            return None
        return snippet["likeCount"], snippet["textOriginal"]
    elif re.search(timestamp, c["snippet"]["textOriginal"]) is not None:
        return c["snippet"]["likeCount"], c["snippet"]["textOriginal"]
