==> src/2b.screen-time.ipynb <==
import numpy as np
import cv2
from matplotlib import pyplot as plt
from IPython.display import Video
import plotly.express as px
import pandas as pd
import time
import os
import pickle
from tqdm import tqdm
from math import floor

WEBSITE = "../../jre-vis/public/"
# TODO How to estimate bandwidth for 2d data?
BANDWIDTH = .1
AVGS_LOCATION = "averages_scale.25.0_color"

def show_frame_from(ep, frame_no):
    video_name = [v for v in video_files if ep.title in v][0]
    cap = cv2.VideoCapture(video_name)
    fps = cap.get(cv2.CAP_PROP_FPS)
    cap.set(cv2.CAP_PROP_POS_FRAMES, frame_no)    
    plt.title(as_timestamp(frame_no, fps))
    _, frame = cap.read()
    plt.imshow(frame)
    cap.release()
    
def as_timestamp(frame_no, fps):
    sec = frame_no / fps
    return f"{floor(sec / 60 / 60)}:{floor(sec / 60) % 60}:{round(sec % 60, 1)}"    

from sklearn.cluster import MeanShift, estimate_bandwidth, KMeans, DBSCAN
from sklearn.neighbors.kde import KernelDensity
from scipy.signal import argrelextrema

def kde_cluster(Xdf, kernel="gaussian", bandwidth=10, plot=False):
    """
    Uses KDE to create clusters out of word vecs
    """
    X = np.array(Xdf['x'])
    X = X.reshape(-1, 1)

    # TODO There is probably bias at the boundaries, should mirror X
    kde = KernelDensity(kernel=kernel, bandwidth=bandwidth).fit(X)
    s = np.linspace(0, np.max(X)*1.5)
    e = kde.score_samples(s.reshape(-1, 1))

    # Reshape back to a 1 by N array
    X = X.reshape(1, -1)

    minima = argrelextrema(e, np.less)[0]
    # Use the linspace to convert back into word indexes
    minima = [s[m] for m in minima]
    # (0, minima 1), (minima 1, minima 2), ... (minima n-1, minima n), (minima n, end)
    minima_pairs = list(zip(np.insert(minima, 0, 0), np.append(minima, s[-1])))

    clusters = [
      np.unique(X[np.logical_and(X >= m1, X < m2)]) for m1, m2 in minima_pairs
    ]

    if plot:
        plt.plot(s, e)
        plt.show()
        print(f"Number of clusters: {len(clusters)}")
        for c in clusters:
            print("\t", len(c), np.unique([int(x) for x in c]))

    return clusters

# Store frame numbers in the clusters instead of the average of the last frame
def frame_idxs_clusters(avg_clusters, df):
    df['cluster'] = np.full_like((len(avg_clusters)), -1)

    for cluster_number, c in enumerate(avg_clusters):
        cluster_cond = (df['x'] >= np.min(c)) & (df['x'] <= np.max(c))
        df.loc[cluster_cond, 'cluster'] = cluster_number
  
    return df

CACHE = "./jre-episodes.pickle"

with open(CACHE, "rb") as f:
    episodes = pickle.load(f)

print(f"Number of loaded episodes: {len(episodes)}")

## Load Averages 

import re
AVERAGES = f"../data/jre/{AVGS_LOCATION}"

for file in os.listdir(AVERAGES):
    matches = re.search(r"-([A-Za-z0-9-_]*).npy", file)
    video_id = matches[1][-11:]
    
    try:
        episode = [e for e in episodes if e.video_id == video_id][0]
        episode.video_averages = np.load(
            f"{AVERAGES}/{file}", allow_pickle=True
        )
    except Exception as e:
        print("Could not load video average data for ", video_id)

Main episodes vs main episodes with video averages

eps = [ep for ep in episodes if ep.video_averages is not None and ep.is_main_episode]
len([ep for ep in episodes if ep.is_main_episode]), len(eps)

### Generate Clusters For Each Episode

def get_frame_count(ep):
    try:
        video_name = [v for v in video_files if ep.title in v][0]
    except IndexError as e:
        print("Could not find video", ep)
        raise e

    cap = cv2.VideoCapture(video_name)
    frames = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
    fps = cap.get(cv2.CAP_PROP_FPS)
    cap.release()
    return frames, fps

def create_clusters(ep, bandwidth):
    averages, total_frames = ep.video_averages
    df = pd.DataFrame({
      'x': averages,
      'n': range(len(averages)),
    })

    avg_clusters = kde_cluster(df, bandwidth=bandwidth, plot=False)
    df = frame_idxs_clusters(avg_clusters, df)

    clusters = {}
    curr_cluster = df['cluster'][0]        
    playhead_start = 0
    skip_amount = total_frames / len(averages)
    playhead = lambda cluster_idx: (cluster_idx * skip_amount) / total_frames

    for i, cluster_num in enumerate(df['cluster']):
        if cluster_num != curr_cluster:
            clusters[curr_cluster] = clusters.get(curr_cluster, []) + [(playhead_start, playhead(i - 1))]
            playhead_start = playhead(i)
            curr_cluster = cluster_num

    return clusters, df

def create_episode_clusters():
    pd.options.mode.chained_assignment = None  # default='warn'

    episode_clusters = []
    episode_clusters_dfs = []
    for ep in tqdm(eps):
        clusters, df = create_clusters(ep,bandwidth=BANDWIDTH)
        episode_clusters.append((ep, clusters))
        episode_clusters_dfs.append((ep, df))
        
    return episode_clusters, episode_clusters_dfs

CACHE = f"./episode-clusters___{AVGS_LOCATION}.pkl"
LOAD = True
# LOAD = False

episode_clusters = []

if LOAD:
    with open(CACHE, "rb") as f:
        episode_clusters, episode_clusters_dfs = pickle.load(f)
else:
    episode_clusters = create_episode_clusters()
    with open(CACHE, "wb") as f:
        pickle.dump(episode_clusters, f)

print("Total ep clusters")
print(len(episode_clusters))

missing_ep_clusters = [e for e,c in episode_clusters if len(c.items()) == 0]

print("Missing ep clusters")
len(missing_ep_clusters), [e.title for e in missing_ep_clusters]

# Generate pie data

pie_rows = []
for ep, cluster_timestamps in episode_clusters:
    data = {}
    for i, t in cluster_timestamps.items():
        data[i] = 0
        for start, end in t:
            data[i] += end - start

    pie_rows.append([ep.video_id, dict(data)])
            
pie_data = pd.DataFrame(pie_rows, columns=["id", "data"])
pie_data.to_csv(WEBSITE + "screen_time.csv")
len(pie_data)

type(episode_clusters[0][1].keys()), episode_clusters[0][1].keys()

# Protobufs are 16mb.. json is 60ish
import json

def key_to_json(data):
    if data is None or isinstance(data, (bool, int, str)):
        return data
    if isinstance(data, (tuple, frozenset)):
        return str(data)
    if isinstance(data, np.int64):
        return int(data)
    raise Exception("Type error", (data, type(data)))

def to_json(data):
    if data is None or isinstance(data, (bool, int, tuple, range, str, list)):
        return data
    if isinstance(data, (set, frozenset)):
        return sorted(data)
    if isinstance(data, dict):
        return {key_to_json(key): to_json(data[key]) for key in data}
    raise TypeError

with open(PROTO_OUT + "screen_time_timelines.json", "w") as f:
    if False:
        json_ep_clusters = json.dumps(to_json({e.video_id:cl for e, cl in episode_clusters}))
        f.write(json_ep_clusters)

# Get pictures for each cluster

import cv2
from glob import glob
import imutils

video_files = list(glob("/Volumes/JRE/jre-bucket/jre/videos/*.mp4"))

BLACK_AND_WHITE = True
RESCALE_AMOUNT = .5

def rescale_frame(frame, amount=.75):
    width = int(frame.shape[1] * amount)
    height = int(frame.shape[0] * amount)
    dim = (width, height)
    return cv2.resize(frame, dim, interpolation =cv2.INTER_AREA)

video_cluster_images = {}
total = 0
for ep, cluster_timestamps in tqdm(episode_clusters):
    video_cluster_images[ep] = {}
    
    try:
        video_name = [v for v in video_files if ep.video_id == v[-15:-4]][0]
    except IndexError as e:
        print("Could not find video", ep)
        continue
    
    # Get a frame for the middle of each cluster timestamp
    for i, t in cluster_timestamps.items():
        cap = cv2.VideoCapture(video_name)
        frames = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
        fps = cap.get(cv2.CAP_PROP_FPS)
       
        # Get the segment in the cluster with the longest time on that shot
        start, end = t[np.argmax([e - s for s, e in t])]
        middle = (end - start) / 2 + start
        # Don't get pictures for items with less than .1% of the total time
        if end - start < .001:            
            continue
        total += 1
        
        # Read frame        
        frame_no = int(frames * middle)
        cap.set(cv2.CAP_PROP_POS_FRAMES, frame_no)
        
        averages, _ = ep.video_averages
        
        ret, frame = cap.read()                

        frame = rescale_frame(frame, amount=RESCALE_AMOUNT)
        if BLACK_AND_WHITE: frame = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
            
        video_cluster_images[ep][i] = {'frame':frame, 'frame_no':frame_no, 'frames':frames, 'fps':fps}        
        
        cap.release()

CACHE = f"./video_cluster_images.pkl"
LOAD_VID = False

if LOAD_VID:
    with open(CACHE, "rb") as f:
        video_cluster_images = pickle.load(f)
else:
    with open(CACHE, "wb") as f:
        pickle.dump(video_cluster_images, f)

missing_video_cluster_images = [k for k,v in video_cluster_images.items() if len(v.items()) == 0]
len(video_cluster_images), len(missing_video_cluster_images)

# Total hours of content

first = lambda x: None if len(list(x.keys())) == 0 else x[list(x.keys())[0]]
total_seconds = np.sum([
    first(info)['frames'] / first(info)['fps'] 
    for x, info in video_cluster_images.items() if first(info) is not None
])
total_seconds / 60 / 60

## Save Images

for ep, clusters in video_cluster_images.items():
    for cluster_n, image in clusters.items():
        cv2.imwrite(f"{WEBSITE}/images/{ep.video_id}.cluster.{cluster_n}.jpg", image['frame'])

### Create timeline data

# protoc --python_out=./ ./screen-time-timeline.proto
# pbf screen-time-timeline.proto --browser > ../../jre-vis/src/lib/proto/screen-time.js
import screen_time_timeline_pb2 as timeline_proto

timelines = timeline_proto.Timelines()

for ep, cluster_timestamps in episode_clusters:
    if len(list(video_cluster_images[ep].values())) == 0:
        continue
        
    timeline = timelines.timelines.add()
    timeline.id = ep.video_id
    timeline.frames = list(video_cluster_images[ep].values())[0]['frames']    
    
    for i, t in cluster_timestamps.items():
        # If it doesn't have a picture, don't bother
        if not i in video_cluster_images[ep]:
            continue
        
        cluster = timeline.clusters.add()
        cluster.id = i
        for start, end in t:
            ts = cluster.timestamps.add()
            ts.start = start
            ts.end = end

PROTO_OUT = "../../jre-vis/public/"

with open(PROTO_OUT + "screen_time_timelines", "wb") as f:
    f.write(timelines.SerializeToString())

## View Images
Image is shown above scatterplot
Its corressponding cluster is highlighted in green
The orange dot indicates the frame it was taken from

from ipywidgets import interact, IntSlider, Dropdown
%matplotlib inline
from ipywidgets.widgets.interaction import show_inline_matplotlib_plots
import ipywidgets as widgets
from IPython.display import clear_output

def show_video_segments():
    def get_title(ep, cluster, timestamp):
        try:
            percentage = pie_data.loc[pie_data['id'] == ep.video_id]['data'].iloc[0][cluster]
        except Exception as e:
            percentage = 0
            print("Could not get percentage for", ep.number)
        return f"{ep.title} -- {cluster} ({round(percentage * 100)}%, {timestamp})"
    
    cluster = choose_cluster.value        
    ep, cluster_images = list(video_cluster_images.items())[ep_slider.value]
    img = cluster_images[cluster]
    
    averages = ep.video_averages[0]
    df = pd.DataFrame({ 'x': averages, 'n': range(len(averages))})
    avg_clusters = kde_cluster(df, bandwidth=BANDWIDTH, plot=False)
    df = frame_idxs_clusters(avg_clusters, df)
    df['color'] = np.where(df['cluster'] == cluster, 'green', 'black')

    with out:
        clear_output()                
        fig, (img_ax, clus_ax) = plt.subplots(2)
        fig.suptitle(get_title(ep, cluster, as_timestamp(img['frame_no'], img['fps'])))        
        
        # Show frame image
        if BLACK_AND_WHITE: img_ax.imshow(img['frame'], cmap='gray')
        else: img_ax.imshow(img['frame'])

        # Show scatter plot of clusters
        # Highlight the frame that was used as orange
        clus_ax.scatter(df['x'], y=df['n'], c=df['color'])        
        cluster_frame = int(img['frame_no'] / img['frames'] * len(df['color']))
        clus_ax.scatter(df.loc[cluster_frame, 'x'], y=df.loc[cluster_frame, 'n'], c='orange')

        show_inline_matplotlib_plots()
        
    
def on_cluster_change(change):
    show_video_segments()
    
def on_ep_change(change):
    new_x = change['new']
    ep, clusters = list(video_cluster_images.items())[new_x]
    choose_cluster.options = clusters.keys()
    choose_cluster.value = list(clusters.keys())[0]    
    show_video_segments()

initial_ep = 0

out = widgets.Output()
display(out)
    
choose_cluster = Dropdown(description='Cluster:')
choose_cluster.observe(on_cluster_change, names='value')

ep_slider = IntSlider(min=0, max=len(video_cluster_images) - 1)
ep_slider.observe(on_ep_change, names='value')
ep_slider.value = initial_ep
on_ep_change({'new': initial_ep})

display(ep_slider, choose_cluster)

# Some episodes fail terribly for screen time
can we find a way to automatically prune them?

from ipywidgets import Dropdown

def show_std(x, e=None, cls=None, ep_clusters=None, logging=False):
    if cls is None or e is None:
        e, cls = episode_clusters_dfs[x]
    
    def log(*e):
        if logging:
            print(*e)

    log(e)
    
    cluster_total = lambda cluster: np.sum([end - start for start, end in cluster])
    # TODO Fix the loading of epcluster so i dont have to do [x][:1]
    stds = []

    for c in np.unique(cls['cluster']):
        try:
            if ep_cluster is None:
                ep_clusters = episode_clusters[x][1]
            c_total = cluster_total(ep_clusters[c])
        except KeyError as e:
            continue

        if c_total < .02:
            continue
            
        xs = cls.where(cls['cluster'] == c)['x']
        ys = cls.where(cls['cluster'] == c)['n']
        xs = np.stack([xs, ys], axis=1)
#         print("-------")
#         print(xs)
#         xs = np.unique(xs)
        xs = xs[~np.isnan(xs)]
#         print("****")
#         print(xs)
        
        log("Cluster", c)
        log(f"total {round(c_total*100)}%")
        log("std", np.std(xs), "\n")
        
        stds.append(np.std(xs))
        
    log()
    stds_sum = np.sum(stds)
    stds_std = np.std(stds)
    log("std total", stds_sum)
    log("std std", stds_std)
    
    return round(stds_sum, 2), round(stds_std, 2)

stds = [(e, show_std(x, False)) for x, e in tqdm(enumerate(eps))]

ones_to_redo = [(stds_std, e) for (e, (std_sum, stds_std)) in stds if stds_std > 15_000]
len(ones_to_redo)
for s, e in ones_to_redo:
    print(e)
    print("old stds std", s)
    
    ep_clusters, cls = create_clusters(e, bandwidth=0.01)
    summed, new_stds_std = show_std(x, e, cls=cls, ep_clusters=ep_clusters, logging=False)
    
    print("NEW ", new_stds_std)
    print("Difference", s - new_stds_std)
    print()
    print()

interact(show_std, x=Dropdown(
    value=9, 
    options=[(f'({std_sum} - {std_total})\t{e.title}', x) for x, (e, (std_sum, std_total)) in tqdm(enumerate(stds))],
))

# Useful for checking out output of kde 
...With different values for the bandwidth

import plotly as py
import plotly.graph_objs as go
from ipywidgets import interact, IntSlider
py.offline.init_notebook_mode(connected = True)

bandwidth = BANDWIDTH
bandwidth = .01
frame_no = 194000

def show_frame_from(ep, frame_no):
    video_name = [v for v in video_files if ep.video_id == v[-15:-4]][0]
    cap = cv2.VideoCapture(video_name)
    fps = cap.get(cv2.CAP_PROP_FPS)
    cap.set(cv2.CAP_PROP_POS_FRAMES, frame_no)    
    plt.title(as_timestamp(frame_no, fps))
    _, frame = cap.read()
    plt.imshow(frame)
    cap.release()

def show_frame_averages(x):
    df = pd.DataFrame({
      'x': eps[x].video_averages[0],
      'n': range(len(eps[x].video_averages[0])),
    })
    
    avg_clusters = kde_cluster(df, plot=False, bandwidth=bandwidth)

    df = frame_idxs_clusters(avg_clusters, df)
    # For a discrete color scale
    df['cluster'] = df['cluster'].astype(str)
    
    
#     show_frame_from(eps[x], frame_no)

    title = f"{str(eps[x])} {len(df['x'])}"
    fig = px.scatter(df, x='x', color='cluster', title=title)
    fig.show()
        

# interact(show_frame_averages, x=IntSlider(value=9, min=0, max=len(eps) - 1))
interact(show_frame_averages, x=Dropdown(
    value=9, 
    options=[(e.title, x) for x, e in enumerate(eps)],
))

# TODO Need some way to flag large deviations in std for clusters
