==> src/3.missing-data.ipynb <==
import pandas as pd
import pickle
import numpy as np
from tqdm import tqdm
import json
from glob import glob

WEBSITE = "../data/jre/website/"

def sort_newest(eps):
    return sorted(eps, key=lambda ep: ep.number, reverse=True)

CACHE = "./jre-episodes.pickle"

with open(CACHE, "rb") as f:
    episodes = pickle.load(f)

print(f"Number of loaded episodes: {len(episodes)}")

main_eps = sort_newest([e for e in episodes if e.is_main_episode])
print("Total main episodes", len(main_eps))
print("Most recent main episode", main_eps[0])

### Missing Main Video Data (uploads.json)

### Missing Captions

# Screen Time
AVGS_LOCATION = "averages_scale.25.0_color"
video_files = list(glob("/Volumes/JRE/jre-bucket/jre/videos/*.mp4"))
video_files_ids = [v[-15:-4] for v in video_files]
downloaded_main = [e for e in main_eps if e.video_id in video_files_ids]
missing_videos = sort_newest(list(set(main_eps) - set(downloaded_main)))

print("Total downloaded videos", len(video_files))
print("Downloaded videos that are main episodes", len(downloaded_main))
print("Missing eps", len(missing_videos))
missing_videos

### Missing Videos
### Missing Video Cluster Images

### Missing Guest Data
set([e.guests for e in episodes])
